requirements.txt
aiodns==3.1.1
aiohttp==3.9.3
aiosignal==1.3.1
arrow==1.3.0
attrs==23.2.0
Brotli==1.1.0
cffi==1.16.0
discord-timestamps @ git+https://github.com/zachlagden/discord-timestamps.git@c5152dd6ed0431a372258db61da4f52ef0cdcedf
discord.py==2.3.2
dnspython==2.6.1
frozenlist==1.4.1
fuzzywuzzy==0.18.0
idna==3.6
Levenshtein==0.25.1
multidict==6.0.5
orjson==3.9.15
pycares==4.4.0
pycparser==2.21
pymongo==4.6.2
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
rapidfuzz==3.8.1
six==1.16.0
termcolor==2.4.0
types-python-dateutil==2.9.0.20240316
yarl==1.9.4
